+++ dot_man/__init__.py
@@ -1,4 +1,4 @@
 """dot-man: Dotfile manager with git-powered branching."""
 
-__version__ = "0.2.0"
+__version__ = "0.3.0"
 __author__ = "Bishoy Ehab"